+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/utropicmedia/storj-influxdb/cmd"
+import "github.com/storj-thirdparty/connector-influxdb/cmd"
 
 func main() {
 	cmd.Execute()